requirements.txt
aiogram==3.2.0
aiohttp-retry==2.8.3
aiohttp[speedups]==3.9.1
aiolimiter==1.1.0
anyio==4.2.0
asyncclick==8.1.3.4
asyncpg==0.29.0
inflect==7.0.0
python-dateutil==2.8.2
SQLAlchemy==2.0.23